feat(phase-19/10): expand multi-agent-software-team into multi-file TypeScript project
