+++ phases/19-capstone-projects/10-multi-agent-software-team/code/ts/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "multi-agent-software-team",
+  "private": true,
+  "type": "module",
+  "version": "0.1.0",
+  "description": "Multi-file TypeScript skeleton for the multi-agent software team capstone.",
+  "scripts": {
+    "start": "tsx src/index.ts",
+    "typecheck": "tsc --noEmit",
+    "test": "tsx --test tests/*.test.ts"
+  },
+  "dependencies": {
+    "hono": "^4.6.0",
+    "zod": "^3.23.0"
+  },
+  "devDependencies": {
+    "@types/node": "^20.14.0",
+    "tsx": "^4.19.0",
+    "typescript": "^5.7.0"
+  }
+}
